moved filter function to it's own file for testing
